MasterVersionType moved to avalon core's lib to be accessable in pipeline.py
