+++ avalon/tools/delegates.py
@@ -3,6 +3,7 @@ import numbers
 from ..vendor.Qt import QtWidgets, QtCore, QtGui
 from .. import io
 from . import lib
+from ..lib import MasterVersionType
 from .models import TreeModel
 
 try:
@@ -19,7 +20,7 @@ class VersionDelegate(QtWidgets.QStyledItemDelegate):
     lock = False
 
     def displayText(self, value, locale):
-        if isinstance(value, lib.MasterVersionType):
+        if isinstance(value, MasterVersionType):
             return lib.format_version(value, True)
         assert isinstance(value, numbers.Integral), (
             "Version is not integer. \"{}\" {}".format(value, str(type(value)))
@@ -98,10 +99,10 @@ class VersionDelegate(QtWidgets.QStyledItemDelegate):
         if master_version and doc_for_master_version:
             version_name = doc_for_master_version["name"]
             label = lib.format_version(version_name, True)
-            if isinstance(value, lib.MasterVersionType):
+            if isinstance(value, MasterVersionType):
                 index = len(versions)
             master_version["data"] = doc_for_master_version["data"]
-            master_version["name"] = lib.MasterVersionType(version_name)
+            master_version["name"] = MasterVersionType(version_name)
 
             item = QtGui.QStandardItem(label)
             item.setBackground(QtGui.QColor(60, 60, 60))
