+++ avalon/tools/lib.py
@@ -1,35 +1,15 @@
 import os
 import sys
 import contextlib
-import numbers
 
 from .. import io, api, style
 from ..vendor import qtawesome
 
-from ..vendor.Qt import QtWidgets, QtCore, QtGui
+from ..vendor.Qt import QtWidgets, QtCore
 
 self = sys.modules[__name__]
 self._jobs = dict()
 self._path = os.path.dirname(__file__)
-
-
-class MasterVersionType(object):
-    def __init__(self, version):
-        assert isinstance(version, numbers.Integral), (
-            "Version is not integer. \"{}\" {}".format(
-                version, str(type(version))
-            )
-        )
-        self.version = version
-
-    def __str__(self):
-        return str(self.version)
-
-    def __int__(self):
-        return int(self.version)
-
-    def __format__(self, format_spec):
-        return self.version.__format__(format_spec)
 
 
 def format_version(value, master_version=False):
@@ -37,7 +17,7 @@ def format_version(value, master_version=False):
     label = "v{0:03d}".format(value)
     if not master_version:
         return label
-    return "<{}>".format(label)
+    return "[{}]".format(label)
 
 
 def resource(*path):
